refactor(blorktools): rig methods to manager
